jobjitsu_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
python-dotenv==1.0.0
supabase==2.0.0
google-generativeai==0.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
